==> analysis/2_data_descriptives.Rmd <==
---
title: "Target Absent Stopping Rules descriptives"
output: 
  pdf_document:
  toc: true
  toc_depth: 2
fig_height: 4
fig_width: 6
---

```{r setup, include=FALSE}
library(tidyverse)
library(rethinking)
load("scratch/processed_data.rda")
knitr::opts_chunk$set(fig.width=12, fig.height=6.5) 
```

``` {r round the difficulty numbers to one decimal place, include=FALSE}
df$difficulty <- round(df$difficulty, digits = 1)
df$p_difficulty <- round(df$p_difficulty, digits = 1)
```

```{r introduce TATP variable}
df$TATP <- 0 

# Say when it's TA and when it's TP 
df$TATP[df$key == "l"] <- "TP"
df$TATP[df$key == "r"] <- "TA"

```

```{r remove the Essex SIBL data, include=FALSE}
toberemoved <- which(df$site == "Essex" & df$group == "SIBL")
df <- df[-toberemoved,]

rm(toberemoved)
```


```{r calculate no. trials, include=FALSE}
total_observations <- length(df$rt)

# get average number of trials 
temp <- group_by(df, participant,block_type,block,group)
number_of_trials <- summarise(temp, num_trials = length(rt))
rm(temp)
```

``` {r display numbers for no. trials}
total_observations

trials_per_participant = mean(number_of_trials$num_trials)
trials_per_participant
```

```{r remove rows containg NA, include=FALSE}
df <- df[complete.cases(df),]

total_observations_nar <- length(df$rt)

temp <- group_by(df, participant,block_type,block,group)
number_of_trials <- summarise(temp, num_trials = length(rt))
rm(temp)

trials_per_participant_nar <- mean(number_of_trials$num_trials)

percent_remaining <- round((total_observations_nar/total_observations)*100)
```



Data collect from Essex for the SIBL group was removed as this data had some issues with block length and also the presence of a "difficulty" level that wasn't included in the study as it was deemed too difficult. 

Each participant carried out 50 trials for each difficulty and in both conditions they took part in. Some responses were removed as participants pressed the wrong key.

The rt values in these were replaced with NA as the wrong key had been pressed. Also, as we were interested in looking at the influence of previous rt's we also removed the first trial from each block as this trial had no previous trial information. 

Prior to removing these trials, there were `r total_observations` trials across participants. After removing cases with NA values, we were left with `r total_observations_nar` trials across all participants (about `r percent_remaining`% of the data). On average, this left participants with `r round(trials_per_participant_nar)` trials per block, condition, and difficulty.


```{r show new numbers}
total_observations_nar
trials_per_participant_nar
```
We can also plot how many participants carried out "x" amount of trials per block. This plot can be seen below as well as a plot containing information about how many trials in total were carried out and how many participants had this many total trials. 

``` {r make trials per participant:block graph, include=FALSE}
trials_per_block_plt = ggplot(number_of_trials, aes(num_trials))
trials_per_block_plt = trials_per_block_plt + geom_bar()
trials_per_block_plt = trials_per_block_plt + xlab("Number of trials")
trials_per_block_plt = trials_per_block_plt + ylab("No. participants")

```

```{r make trials per participant total graph, include = FALSE}
temp <- group_by(df, participant,block_type,group)
number_of_trials_2 <- summarise(temp, num_trials = length(rt))
rm(temp)

trials_total_plt = ggplot(number_of_trials_2, aes(num_trials))
trials_total_plt = trials_total_plt + geom_bar()
trials_total_plt = trials_total_plt + xlab("Number of trials")
trials_total_plt = trials_total_plt + ylab("No. participants")

```

```{r show graphs, echo = FALSE}
trials_per_block_plt
trials_total_plt
```

``` {r get summary stats, include=FALSE}
# so mean and median rt by group, block type, difficulty, and Ta vs Tp
temp = group_by(df, group, block_type, difficulty, targ_pr)
desc_stats = summarise(temp, mean_rt = mean(rt),
                             median_rt = median(rt),
                             sdev = sd(rt),
                             N = length(rt), 
                             se = sdev/sqrt(N),
                             upper_mean = mean_rt + se,
                             lower_mean = mean_rt - se,
                             upper_median = median_rt + se,
                             lower_median = median_rt - se)
rm(temp)

# add in better titles for plots
# initialise empty column 
desc_stats$group2 <- NA
desc_stats$targ_pr2 <- NA

# fill it in
desc_stats$group2[desc_stats$group == "RABL"] <- "Random/Blocked"
desc_stats$group2[desc_stats$group == "SIBL"] <- "Sinewave/Blocked"

desc_stats$targ_pr2[desc_stats$targ_pr == 1] <- "Present"
desc_stats$targ_pr2[desc_stats$targ_pr == 0] <- "Absent"


mean_rt_plot = ggplot(desc_stats, aes(difficulty, mean_rt, colour = block_type))
mean_rt_plot = mean_rt_plot + geom_point() + geom_line()
mean_rt_plot = mean_rt_plot + geom_errorbar(aes(ymin = lower_mean, ymax = upper_mean))
mean_rt_plot = mean_rt_plot + labs(title = "Mean response times accross difficulty for target present and absent trials",
                                   x = "Difficulty", y = "Response Times (seconds)",
                                   colour = "Block Type")
mean_rt_plot = mean_rt_plot + facet_grid(targ_pr2~group2)

median_rt_plot = ggplot(desc_stats, aes(difficulty, median_rt, colour = block_type))
median_rt_plot = median_rt_plot + geom_point()
median_rt_plot = median_rt_plot + geom_errorbar(aes(ymin = lower_median, ymax = upper_median))
median_rt_plot = median_rt_plot + facet_grid(group~targ_pr)

```


Below are plots of the mean and median reaction times separated by group (either RABL or SIBL) and also whether the target was present or absent (1 or 0 respectively). The first graph is the mean reaction times with standard error bars, the second is of the median reaction times. Along the x-axis is the "difficulty" of the trial. Higher numbers mean more variance in the distractors and therefore a more difficult trial. The number relates to how much the distractor line segments could vary about an orientation that was orthogonal to the target item.
```{r display the mean and median rt graphs, echo = FALSE}
mean_rt_plot
median_rt_plot
```

```{r make the same plots but for accuracy this time, include=FALSE}
temp = group_by(df, group, block_type, difficulty, targ_pr)
desc_stats_acc = summarise(temp, mean_correct = mean(correct),
                                 median_correct = median(correct),
                                 sdev = sd(correct),
                                 N = length(correct), 
                                 se = sdev/sqrt(N),
                                 upper_mean = mean_correct + se,
                                 lower_mean = mean_correct - se,
                                 upper_median = median_correct + se,
                                 lower_median = median_correct - se)
rm(temp)

# add in better titles for plots
# initialise empty column 
desc_stats_acc$group2 <- NA
desc_stats_acc$targ_pr2 <- NA

# fill it in
desc_stats_acc$group2[desc_stats_acc$group == "RABL"] <- "Random/Blocked"
desc_stats_acc$group2[desc_stats_acc$group == "SIBL"] <- "Sinewave/Blocked"

desc_stats_acc$targ_pr2[desc_stats_acc$targ_pr == 1] <- "Present"
desc_stats_acc$targ_pr2[desc_stats_acc$targ_pr == 0] <- "Absent"


mean_acc_plot = ggplot(desc_stats_acc, aes(difficulty, mean_correct, colour = block_type))
mean_acc_plot = mean_acc_plot + geom_point() + geom_line()
mean_acc_plot = mean_acc_plot + geom_errorbar(aes(ymin = lower_mean, ymax = upper_mean))
mean_acc_plot = mean_acc_plot + labs(x = "Difficulty", y = "Proportion correct responses", title = "Average Accuracy in Target present and Target Absent trials across difficulty.", colour = "Block Type")
mean_acc_plot = mean_acc_plot + facet_grid(targ_pr2~group2)

```

Below are the plots of mean accuracy over different difficulties, block types, and target presence. 
```{r display the mean accuracy graphs, echo = FALSE}
mean_acc_plot
```

```{r make some density plots, include=FALSE}

# make separate density plot data frame 
dense_data <- df

# make targ_pr a factor 
dense_data$targ_pr <- as.factor(dense_data$targ_pr)

# make difficulty a factor just for this 
dense_data$difficulty <- as.factor(dense_data$difficulty)

density_plot_RABL = ggplot(dense_data[dense_data$group == "RABL",], aes(rt, colour = targ_pr))
density_plot_RABL = density_plot_RABL + geom_density(alpha = 0.1)
density_plot_RABL = density_plot_RABL + facet_wrap(~block_type)
density_plot_RABL = density_plot_RABL + ggtitle("RABL group")


density_plot_SIBL = ggplot(dense_data[dense_data$group == "SIBL",], aes(rt, colour = targ_pr))
density_plot_SIBL = density_plot_SIBL + geom_density(alpha = 0.1)
density_plot_SIBL = density_plot_SIBL + facet_wrap(~block_type)
density_plot_SIBL = density_plot_SIBL + ggtitle("SIBL group")



```

These are the density plots for Target absent and target present responses across groups and the block type. 

```{r display the density plot, echo = FALSE}
density_plot_RABL
density_plot_SIBL
```

```{r density plots with difficulty as a factor, include=FALSE}

density_plot_RABL_diff = ggplot(dense_data[dense_data$group == "RABL",], aes(rt, colour = difficulty))
density_plot_RABL_diff = density_plot_RABL_diff + geom_density(alpha = 0.1)
density_plot_RABL_diff = density_plot_RABL_diff + facet_grid(targ_pr~block_type)
density_plot_RABL_diff = density_plot_RABL_diff + ggtitle("RABL group")


density_plot_SIBL_diff = ggplot(dense_data[dense_data$group == "SIBL",], aes(rt, colour = difficulty))
density_plot_SIBL_diff = density_plot_SIBL_diff + geom_density(alpha = 0.1)
density_plot_SIBL_diff = density_plot_SIBL_diff + facet_grid(targ_pr~block_type)
density_plot_SIBL_diff = density_plot_SIBL_diff + ggtitle("SIBL group")

```


These are the plots when we distinguish across the different difficulty levels. For the purpose of these graphs, difficulty was treated as a "discrete" factor, but for the modelling this should be a continuous variable instead. I created a different data frame for these graphs so the main data frame (df) should remain unchanged. 

``` {r display the density plots with difficulty as a factor, echo = FALSE}
density_plot_RABL_diff
density_plot_SIBL_diff
```

We took a look at trial number (across blocks) to look at whether there were any practice effects that may need to be taken account of when creating a model of the data. 

```{r Looking at trial number, echo = FALSE}
temp = group_by(df, difficulty, block_type, targ_pr, trial)
trial_desc_stats = summarise(temp, mean_rt = mean(rt),
                             median_rt = median(rt),
                             sdev = sd(rt),
                             N = length(rt), 
                             se = sdev/sqrt(N),
                             upper_mean = mean_rt + se,
                             lower_mean = mean_rt - se,
                             upper_median = median_rt + se,
                             lower_median = median_rt - se,
                             # log mean
                             log_mean_rt = mean(log2(rt)),
                             log_sdev = sd(log2(rt)),
                             log_se = log_sdev/sqrt(N),
                             log_mean_upper = log_mean_rt + log_se,
                             log_mean_lower = log_mean_rt - log_se,
                             # log median
                             log_median_rt = median(log2(rt)),
                             log_median_upper = log_median_rt + log_se,
                             log_median_lower = log_median_rt - log_se)
# tidy
rm(temp)

trial_plt = ggplot(trial_desc_stats, aes(trial, log_median_rt, colour = difficulty))
trial_plt = trial_plt + geom_point()
trial_plt = trial_plt + geom_smooth(method = lm)
trial_plt = trial_plt + facet_grid(targ_pr~block_type)

trial_plt
```


# Modelling data creation 
Below is a record of what happened for the modelling of the data. Every step taken to produce each model and the results of each can be seen here. 

Initially, we centred all contniuos predictors according to the grand mean 
```{r centreing the predictors}
# centre p_rt
df$c_p_rt <- df$p_rt - mean(df$p_rt)
```

Difficulty was scaled by dividing everything by pi. This means that 0 is the easiest difficulty (no variance in the distractors) and 1 is virtually impossible (the distractors vary through 360$deg;)

Difficulty was also changed to be called theta

``` {r scale difficulty}
# change difficulty name
#colnames(df)[7] <- "theta"

#rescale difficulty for the models
df$theta <- round(df$difficulty/pi, digits = 3) 

# rescale p_difficulty
df$p_theta <- round(df$p_difficulty/pi, digits = 3)

# create a centred version 
df$theta_c <- df$theta - mean(df$theta)
```

A new variable was created using current and previous theta. This is to reflect how much change there was from trial to trial in terms of the difficulty. Negative numbers mean the current trail is easier than the previous, positive means the current trial is harder than the previous. 

``` {r Get comparison of current trial to previous}
df$change_th <- df$theta - df$p_theta
```

We created dummy variables for the contrasts between block_type. Blocked being 0

```{r Create dummy variables for block_type}
df$isra <- ifelse(df$block_type == "random", 1,0) 

df$issi <- ifelse(df$block_type == "sinewave",1,0)
```

Need to fix participant so it is a continuous string of numbers, otherwise R gets confused

```{r redo participant coding}
df$participant <- as.factor(df$participant)
df$participant <- as.numeric(df$participant)
```

We also created a dataset just to look at correct trials. 

```{r make only correct trials dataset}
df_correct_only <- df[df$correct == 1,]
```

``` {r create save file, include=FALSE}
# remove everything but the df 
rm(list=ls()[!(ls() %in% c("df","df_correct_only"))])

# Save only the new data frame
save(df, df_correct_only, file = "scratch/processed_data_nar.rda")
```

``` {r working on density plots for TA trials as a probability of TP rts, include = FALSE}

prob_data <- function(df, p, th){
  # choose one participant
  d <- df[df$participant == p,]
  
  # choose difficulty
  d <- d[d$difficulty == th,]
  
  # get log of rt 
  d$logrt <- log(d$rt)
  
  # get mean & sd logrt for TP trials
  mu <- mean(d$logrt[d$TATP == "TP"])
  sig <- sd(d$logrt[d$TATP == "TP"])
  
  # pnorm on logrt
  d$pnormRT <- pnorm(d$logrt, mu, sig)
  
  # add into data frame 
  return(d) 
}

# set up the data frame for this part
# We only want correct trials, so we should use that for now
# though, we might want to include incorrect trials for the TA part and separate on that?
df_pnorm <- df_correct_only

# before running this, we need to make sure there are empty columns to add into
df_pnorm$logrt = NA
df_pnorm$pnormRT = NA

# now create the loop
for(th in unique(df_pnorm$difficulty)){
  for(p in unique(df_pnorm$participant)){
    d <- prob_data(df_pnorm, p, th)
    
    # add into datafram 
    df_pnorm[which(df_pnorm$participant == p & df_pnorm$difficulty == th),] <- d
  }
  # create filename
  filename <- paste("scratch/plots/diff_", th, "_probability_plots.pdf", sep = "")
  
  # make the plot 
  plt = ggplot(df_pnorm[df_pnorm$difficulty == th,], aes(pnormRT, colour = TATP))
  plt = plt + geom_density(alpha = 1)
  plt = plt + scale_y_continuous(limits = c(0,9))
  plt = plt + facet_wrap(~participant)
  plt
  ggsave(filename, height = 10, width = 8)
}
```

# Working on TA trials only

As target absent judegemnts were our primary interest, we decided to create a dataset of only "Target Absent" judgements. The thought process is that making a Target Present judgement is due to detecting the target, whereas we are interested in how people make the decision that the target is not there.

``` {r makes block type an index variable, include = FALSE}
df$block_type_id <- coerce_index(df$block_type)
```

```{r Create dataset of only TA judgement trials}
# subset the data 
df_TA <- df[df$TATP == "TA",]

#tidy up
rm(df, df_correct_only)

# save this 
save(df_TA, file = "scratch/processed_data_nar_TA.rda")
```

Create some box plots of participants RTs for TA trials only

``` {r Box plots of just TA data}
df_TA_bxplt <- df_TA
df_TA_bxplt$participant <- as.factor(df_TA_bxplt$participant)

boxplt <- ggplot(df_TA_bxplt, aes(x = participant, y = log2(rt)))
boxplt <- boxplt + geom_boxplot()

boxplt
```


# Notes

## Internal numbering
Rethinking's internal numbering is like 1,10,11,etc. so this needs to be considered when extracting information from the models. For example, when looking at the intercepts, a_p[2] actually corresponds to participant 10, not 2. This fix has been added to the modelling scripts where this is important. 

## Dummy Variable coding 
This was used when looking at block type. For all models, "isra" refers to trial in the random condition, "issi" refers to trials in the sine wave condition.

``` {r get log values for intercept etc., include = FALSE}
anna_a_easy <- 1.75
anna_a_easy_sd <- 0.13

log_a_easy <- log(anna_a_easy)
log_a_easy_sd <- log(anna_a_easy_sd)

anna_a_hard <- 3.94
anna_a_hard_sd <- 2.19

log_a_hard <- log(anna_a_hard)
log_a_hard_sd <- log(anna_a_hard_sd)

```

## Informed priors
We can use the data from Anna's study to get a feel for what the intercept and slope should be in our model. 

Remembering that we are operating on a log scale, we need to change the values so they are appropriate for the model.

Anna's data suggests that at the easiest difficulty, the mean was `r anna_a_easy` (sd = `r anna_a_easy_sd`) for the easiest trials. For the hard trials it was `r anna_a_hard` (SD = `r anna_a_hard_sd`). 

For our model, we should set the intercept value (or $a$ in the model) to be the log(`r anna_a_easy`)  which is `r round(log_a_easy, digits = 3)`. As for the $\sigma$ component of $a$ ~ dnorm(`r round(log_a_easy, digits = 3)`, $\sigma$), this can be set to 1 as this gives a resonable range for RTs.

## PRT parameter
This doesn't seem to help the fit of the model when added as an interaction with block_type. However, this could be because the changes in what the previous trial's difficulty was, so we should think about how to do this? As a categorical variable (i.e. the current trial is easier than the previous etc?), or as a continuous variable (i.e. the difference between current theta and previous theta?)
